requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
oauthlib==3.1.1
PySocks==1.7.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
tweepy==3.10.0
urllib3==1.26.6
python-decouple==3.4
